Fix deprecation warnings for procedure syntax
This syntax for side-effect functions has been deemed a misfeature of
the language that adds complexity for little value, so it’s being
removed in the future. The alternative is to simply add the = sign, or
additionally add the explicit `Unit` return type annotation which is
always advised for methods that are public interface.

It is also a strongly encouraged style convention in Scala to always
call pure side-effect functions with empty parentheses, to make them
visually distinct from value access.