Change CSV parsing to kantan.csv instead of ad hoc
This cleans up a lot of the messy fiddling like dealing with commas
present within quoted fields, letting a full parser implementation deal
with those for us. It can decode to case classes easily, and seems
noticeably faster.

On the other hand this change doesn’t deal with replacing backslashes as
we were doing. This wasn’t ideal anyway, they mainly occur for fields
encoded as "\N" which are meant to represent null values. We ought to
encode nullable fields as `Option[T]` with "\N" values converted to
`None`, but we can look into that when a need arises.

Some noteworthy docs for quickly understanding kantan.csv:

  - https://nrinaudo.github.io/kantan.csv/tut/rows_as_case_classes.html

  - https://nrinaudo.github.io/kantan.csv/tut/data_as_collection.html