Finish making RouteMap completely asynchronous
No more blocking Await usage. This was so effective, in fact, that we
started to blow the queue limit that Slick imposes for actions on its
thread pool, since we're executing queries while traversing graphs which
is almost surely exponential O:-)

The hacky workaround, for now, is to set that queue to be unbounded...
Ideally for real-world usage there would be memoization when building
graph structures, or probably a graph DB. Ironically we're worse off now
for adding the relational DB than previously when doing everything in
memory, but the main point for the DB was mainly to persist a file
instead of loading from CSV on every run, which at this point we're
still doing.